requirements.txt
pytest==7.4.2
allure-pytest==2.13.2
python==3.10.12
pytest-xdist==3.5.0
Appium-Python-Client==4.0.0